Updated version
I'm adding all files again, I changed a lot in code.

I fixed the file saving and added constructors for maps from files. I added the ability to search for guests by room and guest code. I added the ability to enter two-part addresses. I removed the unnecessary option of adding the address as several sub-elements of a string vector. It is enough to add it as a whole, as no one needs to search only for the postal code. I also fixed multiple errors that came up during the revisions.